Update flask and local instructions
Flask needed to be upgraded to be compatible with markupsafe (which in
turn needed to be updated to be compatible with wekzeug updated by
dependabot). While testing locally I noticed that vercel CLI no longer
properly supports rewrites for backends, even though they still work in
production. I updated the README to use next dev and flask debug server
directly, and made some updates to the flask app to be compatible with
the next dev proxy.